refactor(variants): restructure Text and Link components
