+++ packages/core/variants/src/components/link-variants/index.ts
@@ -1,0 +1,2 @@
+export { default as linkVariants } from "./link-variants";
+export type { LinkVariantProps } from "./link-variants";
