--- packages/core/variants/src/components/link/index.ts
@@ -1,2 +1,0 @@
-export { default as Link } from "./link";
-export type { LinkVariantProps } from "./link";
