+++ packages/core/variants/src/components/text-variants/index.ts
@@ -1,0 +1,2 @@
+export { default as textVariants } from "./text-variants";
+export type { TextVariantProps } from "./text-variants";
