--- packages/core/variants/src/components/text/index.ts
@@ -1,2 +1,0 @@
-export { default as Text } from "./text";
-export type * from "./text";
