+++ packages/react/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Text } from "./text";
-export type * from "./text";
+export type { TextProps } from "./text";
 
 export { default as Link } from "./link";
-export type * from "./link";
+export type { LinkProps } from "./link";
